Move Window Types to seperate files
