+++ src/openloco/openloco.vcxproj
@@ -108,7 +108,8 @@
     <ClCompile Include="windows\map.cpp" />
     <ClCompile Include="windows\MessageWindow.cpp" />
     <ClCompile Include="windows\music_selection.cpp" />
-    <ClCompile Include="windows\NewsWindow.cpp" />
+    <ClCompile Include="windows\News\News.cpp" />
+    <ClCompile Include="windows\News\Ticker.cpp" />
     <ClCompile Include="windows\ObjectSelectionWindow.cpp" />
     <ClCompile Include="windows\options.cpp" />
     <ClCompile Include="windows\PlayerInfoPanel.cpp" />
@@ -242,6 +243,8 @@
     <ClInclude Include="window.h" />
     <ClInclude Include="windows\construction\Construction.h" />
     <ClInclude Include="message.h" />
+    <ClInclude Include="windows\News\Common.h" />
+    <ClInclude Include="windows\News\News.h" />
     <ClInclude Include="windows\toolbar_top_common.h" />
   </ItemGroup>
   <ItemGroup>
